+++ common/src/main/java/com/macuguita/lib/platform/registry/GuitaRegistries.java
@@ -1,0 +1,17 @@
+package com.macuguita.lib.platform.registry;
+
+import dev.architectury.injectables.annotations.ExpectPlatform;
+import net.minecraft.registry.Registry;
+import org.apache.commons.lang3.NotImplementedException;
+
+public class GuitaRegistries {
+
+    public static <T> GuitaRegistry<T> create(GuitaRegistry<T> parent) {
+        return new GuitaRegistryChild<>(parent);
+    }
+
+    @ExpectPlatform
+    public static <T> GuitaRegistry<T> create(Registry<T> registry, String id) {
+        throw new NotImplementedException();
+    }
+}